Move createEventService to EventApiFactory
