+++ data/api-events/src/main/kotlin/com/uhc/api/events/di/ApiEventsModule.kt
@@ -1,6 +1,6 @@
 package com.uhc.api.events.di
 
-import com.uhc.api.events.EventApi
+import com.uhc.api.events.EventApiFactory
 import com.uhc.api.events.interceptor.RemoteRequestInterceptor
 import com.uhc.lib.network.utils.interceptors.ApiErrorInterceptor
 import org.koin.core.module.dsl.factoryOf
@@ -11,5 +11,5 @@ val apiEventsModule = module {
     factoryOf(::ApiErrorInterceptor)
     factoryOf(::RemoteRequestInterceptor)
 
-    singleOf(EventApi::createEventService)
+    singleOf(EventApiFactory::createEventService)
 }
